refactor: Redid the schema file to be more descriptive
Previous schema file was generated using mysql_dump which has its own way
of generating the file so everything was in an order that was hard to
read. Redid the schema file to organise the create statements in a more
logical order and added comments to each section for better readability.